# source/asr-candidate/docs/api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "云蝠 Voice Runtime 管理与观测接口",
    "version": "1.13.0",
    "description": "云蝠 Voice Runtime 管理API：保留FreeSWITCH迁移对照，优先面向语音智能体媒体链路。1.10新增真实G711处理能力状态、显式隔离压测模式和完整内容/采样验收。未实现能力保持明确边界，不宣称完整FreeSWITCH或生产万路认证。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:9080",
      "description": "本机示例；实际由active.admin.listen决定。"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/tests": {
      "get": {
        "operationId": "getTests",
        "summary": "读取测试能力与任务列表",
        "description": "只读返回当前控制进程的测试能力、唯一运行任务和最近20个终态摘要。当前任务附最近至多300条成功采样，历史摘要的samples为空数组；所有记录只在内存，重启后消失。档位不是实测通过容量，实际FD、隔离端口和二进制检查在显式创建后执行。",
        "tags": [
          "压力与电话测试"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "当前能力和内存任务索引；current可能为null，history最多20个摘要，每个摘要samples为空数组。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestOverview"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "x-freeswitch": {
          "interface": "无同名HTTP测试作业入口；originate/status仅为原版呼叫与观测",
          "relationship": "analogous_not_wire_compatible",
          "difference": "作业由本项目启动本机隔离实例、模拟两端并验证真实RTP，不兼容ESL/bgapi，不认证FreeSWITCH运行能力。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        }
      },
      "post": {
        "operationId": "startTest",
        "summary": "启动压力或单路媒体测试",
        "description": "仅创建本机独立实例，不接受任意可执行文件、主机或路径。202只表示受理，后续资源预检仍可能失败。request_id在当前控制进程内去重，最多4096个键；同键同参数返回200，同键异参、有运行任务、服务退出或清理冻结返回409；报告淘汰后原键返回410且不重新执行，键数达上限的新请求返回429。总执行上下文期限T=ceil(concurrency/cps)+duration_seconds+90秒，到期进入终止清理，报告发布时间可晚于T。",
        "tags": [
          "压力与电话测试"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "202": {
            "description": "已原子占用唯一测试槽并异步执行预检；返回完整任务记录，不保证已经启动或能够达到目标并发。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            }
          },
          "200": {
            "description": "当前进程中同request_id、同参数的既有任务；不会重复执行，报告保留范围内可恢复实际状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            }
          },
          "409": {
            "description": "已有测试运行、服务正在退出、资源清理失败导致冻结，或同request_id对应不同参数。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "该幂等键对应的报告已从最近20份内存历史淘汰；键仍保留，明确拒绝重新执行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "模式、并发、CPS、时长或载荷越界。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "description": "JSON类型、未知字段、空正文或拼接JSON无效。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "必须为application/json。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "本控制进程已保存4096个幂等键，拒绝新增任务；既有键仍按去重规则处理，重启后内存记录清空。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名HTTP测试作业入口；originate/status仅为原版呼叫与观测",
          "relationship": "analogous_not_wire_compatible",
          "difference": "作业由本项目启动本机隔离实例、模拟两端并验证真实RTP，不兼容ESL/bgapi，不认证FreeSWITCH运行能力。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequest"
              }
            }
          }
        }
      }
    },
    "/v1/tests/{id}": {
      "get": {
        "operationId": "getTest",
        "summary": "读取指定测试进度",
        "description": "返回保留中的任务阶段、发生器进度、最近至多300条成功采样和清理状态。采样错误不会被填成零值；全程峰值在最终报告的evidence中。历史报告最多20份且只保留在内存，淘汰或重启后返回404。终态字段本身不是生产容量通过证明。",
        "tags": [
          "压力与电话测试"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "创建或列表实际返回的任务ID，不能作为文件路径。",
            "schema": {
              "type": "string"
            },
            "example": "TASK_ID_FROM_CREATE"
          }
        ],
        "responses": {
          "200": {
            "description": "保留中的完整任务；samples最多300条，只包含成功读取的真实状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            }
          },
          "404": {
            "description": "未知或已淘汰任务。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "x-freeswitch": {
          "interface": "无同名HTTP测试作业入口；originate/status仅为原版呼叫与观测",
          "relationship": "analogous_not_wire_compatible",
          "difference": "作业由本项目启动本机隔离实例、模拟两端并验证真实RTP，不兼容ESL/bgapi，不认证FreeSWITCH运行能力。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        }
      }
    },
    "/v1/tests/{id}/stop": {
      "post": {
        "operationId": "stopTest",
        "summary": "停止压力测试或挂断测试小电话",
        "description": "只取消指定任务拥有的控制、Rust媒体和发生器实例，不停止主服务。202表示已请求停止，继续读取cleanup确认资源；200表示指定任务已有终态，重复停止不再次执行。最终报告在有界清理完成后发布，提前停止不形成完整通过结论。记录淘汰或进程重启后返回404。",
        "tags": [
          "压力与电话测试"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "创建或列表实际返回的任务ID，不能作为文件路径。",
            "schema": {
              "type": "string"
            },
            "example": "TASK_ID_FROM_CREATE"
          }
        ],
        "responses": {
          "202": {
            "description": "停止请求已受理，任务进入stopping；资源仍可能处于清理中，需继续读取任务。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            }
          },
          "200": {
            "description": "指定任务已经形成终态报告，返回原任务，不再次停止或重新执行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            }
          },
          "404": {
            "description": "未知或已淘汰任务。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "description": "JSON类型、未知字段、空正文或拼接JSON无效。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "必须为application/json。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "停止正文必须为一个空JSON对象；null或非空对象返回422。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名HTTP测试作业入口；originate/status仅为原版呼叫与观测",
          "relationship": "analogous_not_wire_compatible",
          "difference": "作业由本项目启动本机隔离实例、模拟两端并验证真实RTP，不兼容ESL/bgapi，不认证FreeSWITCH运行能力。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/tests/{id}/report": {
      "get": {
        "operationId": "getTestReport",
        "summary": "下载测试结果与失败证据",
        "description": "成功、失败及停止任务在资源清理结束并发布后均可下载，不可变报告只在本进程内存保留最近20份。运行或清理中返回409，已淘汰或重启后返回404。controller_tail和generator_tail各保留最近64KiB原始日志字节；result未实际生成或未通过有界读取时为null。evidence逐字段记录预检、部分二进制摘要、采样峰值、日志和最后收敛状态；缺失/null不表示通过。",
        "tags": [
          "压力与电话测试"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "创建或列表实际返回的任务ID，不能作为文件路径。",
            "schema": {
              "type": "string"
            },
            "example": "TASK_ID_FROM_CREATE"
          }
        ],
        "responses": {
          "200": {
            "description": "资源清理流程结束后的完整JSON报告，含实际已有的成功或失败证据；通过仍需核对run、result、日志收敛和cleanup。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestReport"
                }
              }
            }
          },
          "404": {
            "description": "未知或已淘汰任务。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "测试仍在运行或清理中，没有可发布的完整报告。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名HTTP测试作业入口；originate/status仅为原版呼叫与观测",
          "relationship": "analogous_not_wire_compatible",
          "difference": "作业由本项目启动本机隔离实例、模拟两端并验证真实RTP，不兼容ESL/bgapi，不认证FreeSWITCH运行能力。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "诊断文本；不要依赖具体错误字串编写控制逻辑。"
          }
        },
        "description": "",
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "TestRun": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "后端生成的任务ID。"
          },
          "request_id": {
            "type": "string",
            "description": "创建幂等键。"
          },
          "mode": {
            "type": "string",
            "description": "测试模式。",
            "enum": [
              "pressure",
              "phone"
            ]
          },
          "status": {
            "type": "string",
            "description": "任务生命周期状态；completed仍需结合result.passed和cleanup。",
            "enum": [
              "preparing",
              "running",
              "stopping",
              "completed",
              "failed",
              "stopped"
            ]
          },
          "phase": {
            "type": "string",
            "description": "编排阶段，与发生器progress.phase分开。",
            "enum": [
              "preflight",
              "setup",
              "media",
              "teardown",
              "finished"
            ]
          },
          "concurrency": {
            "type": "integer",
            "description": "本次目标并发。",
            "minimum": 0
          },
          "cps": {
            "type": "integer",
            "description": "本次拨号速率。",
            "minimum": 0
          },
          "duration_seconds": {
            "type": "integer",
            "description": "配置的媒体发送秒数。",
            "minimum": 0
          },
          "payload": {
            "type": "integer",
            "enum": [
              0,
              8,
              9,
              111,
              18,
              110,
              112,
              113,
              114,
              119,
              120,
              121,
              122,
              123
            ],
            "description": "20ms测试媒体预设PT；名称、采样率和RTP时钟见 /v1/codecs.codec_profiles。真实SIP动态PT不绑定测试预设。"
          },
          "created_at": {
            "type": "string",
            "description": "任务受理时间。"
          },
          "started_at": {
            "type": "string",
            "description": "实际启动时间；未启动时为空。"
          },
          "finished_at": {
            "type": "string",
            "description": "终态时间；运行中为空。"
          },
          "elapsed_seconds": {
            "type": "number",
            "description": "已执行墙钟秒数。",
            "minimum": 0
          },
          "progress": {
            "$ref": "#/components/schemas/TestProgress"
          },
          "samples": {
            "type": "array",
            "description": "最多最近300条真实成功采样，正常运行阶段每500ms尝试读取；收敛阶段可每200ms尝试读取。失败读取不填零样本，history摘要固定为空数组。全程采样峰值单独保存在最终evidence中。",
            "items": {
              "$ref": "#/components/schemas/TestSample"
            },
            "maxItems": 300
          },
          "error": {
            "type": "string",
            "description": "失败或停止原因，成功时为空。"
          },
          "report_available": {
            "type": "boolean",
            "description": "可读取报告；不等于测试通过。"
          },
          "cleanup": {
            "$ref": "#/components/schemas/TestCleanup"
          },
          "failure": {
            "$ref": "#/components/schemas/TestFailure"
          },
          "media_processing": {
            "type": "string",
            "enum": [
              "relay",
              "g711"
            ],
            "description": "relay 同编码 RTP 透传；g711 PCMA/PCMU 双腿真实解码与编码。"
          }
        },
        "required": [
          "id",
          "request_id",
          "mode",
          "status",
          "phase",
          "concurrency",
          "cps",
          "duration_seconds",
          "payload",
          "created_at",
          "started_at",
          "finished_at",
          "elapsed_seconds",
          "progress",
          "samples",
          "error",
          "report_available",
          "cleanup"
        ],
        "additionalProperties": true,
        "description": "当前进程内的一次测试生命周期记录。current及单任务接口提供最近至多300条成功采样；history摘要仍有samples字段但固定为空数组。created_at为受理时间，started_at为发生器成功启动时间，finished_at为清理结束并发布报告的时间；未进入相应阶段时为空字符串。"
      },
...
      "TestOverview": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "当前是否允许创建测试。"
          },
          "disabled_reason": {
            "type": "string",
            "description": "不可创建时的原因。"
          },
          "capabilities": {
            "$ref": "#/components/schemas/TestCapabilities"
          },
          "current": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TestRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "type": "array",
            "description": "当前控制进程内最近至多20个终态任务摘要，按结束时间倒序；每项保留samples字段但值为空数组。报告淘汰后GET详情/报告返回404，同幂等键重试创建返回410。",
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/TestRun"
            }
          }
        },
        "required": [
          "enabled",
          "disabled_reason",
          "capabilities",
          "current",
          "history"
        ],
        "additionalProperties": true,
        "description": "只查询当前控制进程的内存任务状态，不启动预检或进程。最近20份终态报告随控制进程重启丢失；幂等记录最多4096个，达到上限后新请求返回429。capabilities档位是输入上限，不证明当前机器有足够FD、端口或算力。"
      },
...
    },
    "responses": {
      "Forbidden": {
        "description": "Host不是localhost/回环字面量，或写请求Origin/CSRF校验失败。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "parameters": {
      "Host": {
        "name": "Host",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "description": ""
        },
        "description": "客户端通常自动设置。主机部分须为精确localhost或回环IP字面量，自定义域名即便解析到回环也拒绝。"
      },
...
    },
    "securitySchemes": {
      "LocalCSRF": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RustSwitch-CSRF",
        "description": "读取/v1/config.csrf_token；重启更换，仅防跨站写入，不是远程认证。"
      }
    }
  },
...
}